Selection sort
here by the end of inner iteration we get biggest element in the array , then we swap to LUI, so the sorted portion increases from right to left, So LUI decrease ,we repeat this until entire array is sorted. Here we select a greatest element on each iteration and swap so it is called selection sorting.