Create ACCENTURE
Description

There is a fight club in Delhi. There are N fighters. The strength of each fighter is mentioned in an array of sizes. N. In every match two fighters with maximum strength fight each other. Let the strength of fighters be X and Y respectively. If Xy, then both of them lose else if XsY, X wins but his strength becomes X-Y. He again waits for another fight.

At last, at most one fighter will be present. if one is present return his strength Else return 0.

N-the number of fighters

an-An array having the strength of all fighters

Output-The strength of the last fighter if present else

Constraints

0<N-1000

0 artice 10000

Examples

• Input:

6

274181

Output:

1
Explanation:

We combine 7 and 8 to get 1 so the array converts to [2,4,1,1,1]
 then, we combine 2 and 4 to get 2 so the array converts to [2,1,1,1]

then, we combine 2 and 1 to get 1 so the array converts to [1,1,1] then, we combine 1 and 1 to get 0 so the array converts to [1] 
then that's the strength of the last fighter.